requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask==3.1.0
lyricsgenius==3.2.0
python-dotenv==1.0.1
spotipy==2.24.0
